requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-WTF==1.0.0
email-validator==1.1.3
Pillow==8.3.1
python-dotenv==0.19.0

#python -m venv venv
#venv\Scripts\activate
#pip install -r requirements.txt
#python
#from app import app, db
#app.app_context().push()
#db.create_all()
#exit()
#python app.py

#pip install flask flask-sqlalchemy flask-login flask-wtf pillow python-dotenv